minor bugs

+++ libft/ft_strcmp.c
@@ -6,7 +6,7 @@
 /*   By: mde-la-s <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/12/14 13:53:37 by mde-la-s          #+#    #+#             */
-/*   Updated: 2021/12/14 14:16:21 by mde-la-s         ###   ########.fr       */
+/*   Updated: 2022/02/01 14:46:38 by mde-la-s         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,5 +19,5 @@ int	ft_strcmp(char *str1, char *str2)
 	i = 0;
 	while (str1[i] && str2[i] && str1[i] == str2[i])
 		i++;
-	return (str1[i] - str2[i]);
+	return ((int)str1[i] - (int)str2[i]);
 }
